[vmtest] enable --json-summary option
This will cause test_progs* to write a json summary report to disk.
Later, libbpf/ci actions will be able to collect this file and report a
summary to the GH UI. libbpf/ci#93 is the companion diff to take
advantage of this new summary.

While we could to some extend report to the UI directly from within
vmtest, doing it from outside the VM environemt will allow us to have
richer output such as step summary [0].

[0] https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#adding-a-job-summary

Signed-off-by: Manu Bretelle <chantr4@gmail.com>