move set-test-matrix script into a python file
The logic is quite complicated and error prone to maintain as an inline
script.
This change moves the logic to its own python script and add argument
parsing to be able to pass the values of the github owner and
repository.

This will allow on subsequent diffs to add unittesting as well as adding
some small features like making it easy to switch to a testing matrix
(for instance, while iterating on diffs, it can be useful to only
trigger testing on one architecture/one toolchain instead ofhe whole
suite.

Testedwith manually generating test matrix from the command line per
https://gist.github.com/chantra/49c760fcbd103d7095cbc3f34a441120

and confirmed that bpf-rx repo does not generate tests for s390x.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>