[workflow] Add a 'veristat' job
Adds a job to execute veristat on a selected set of object files as a
part of the general test pipeline. The job is executed differently for
pushes and for pull requests:

- For pushes:
  - the kernel is compiled as usual;
  - selftests are compiled as well, this is necessary to produce
    object files that would be used by veristat;
  - veristat is executed inside qemu, performance statistics is
    collected and cached as baseline data.

- For pull requests:
  - veristat is executed inside qemu and performance statistics is
    collected;
  - baseline data is restored from cache;
  - baseline data is compared to current statistics and step summary
    page is formed;
  - if there are some programs that verified successfully in the
    baseline but fail verification in current run, the step is marked
    as failed.

This also adds 'run_veristat' "entrypoint" for run_selftests.sh.

This entrypoint executes veristat on a set of binary files specified
in ./tools/testing/selftests/bpf/veristat.cfg file and saves ouput to
the /command_output/ directory within qemu image. This output is
extracted from the image by run-qemu action.