run set-matrix on self-hosted runners for kernel-patches org
With bpf-rc repo using GH hosted runners, we have seen cases when bpf
actions are waiting on a GH runner. An org is limited to 20 GH runners,
when multiple series are being processed, we max them out and bpf repo,
which predominantly runs on self-hosted runners, is being idle waiting
for a GH runner to free up, perform the action in a couple of seconds
and move on.

This change will run set-matrix on any self-hosted machine free and
ready to take an action. It does not need any specific arch, so let's
pick whatever.
This mean that we will possibly run set-matrix for repo that are not
running off self-hosted, on self-hosted, but that should not be a
problem. in term of resources.

Here the trick is to evaluate the runner to use as part of a github
action expresion so it is executed by github, and not the runner.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>