[get_test_matrix] make script compatible with python <= 3.9
The script would fail on runners because it does not run the latest
python.

The `dict1 | dict2` expression ot added in python 3.9: https://docs.python.org/3/library/stdtypes.html#dict.values

This change replace the expression by a good old for loop.

Tested the generated content did not change with:
```
$ diff /tmp/bpf-rc.out <(GITHUB_OUTPUT=/dev/stdout python3.8 ./ci/scripts/get_test_matrix.py -o kernel-patches -r kernel-patches/bpf-rc)
$ diff /tmp/bpf.out <(GITHUB_OUTPUT=/dev/stdout python3.8 ./ci/scripts/get_test_matrix.py -o kernel-patches -r kernel-patches/bpf)
```

Signed-off-by: Manu Bretelle <chantr4@gmail.com>