GCC BPF: use pre-built release
Running libbpf/ci/build-bpf-gcc action builds GCC from latest source
snapshot and caches the output with actions/cache.

Because of actions/cache rules, caches between pull requests are
independent. This results in cache being full of identical GCC builds,
which interferes with caching actually useful things.

To avoid this, and also to not re-build GCC all the time, download a
pre-built GCC and use it. The "build release" job is set up in an
independent repository, publishing artifacts as a release tag.

Signed-off-by: Ihor Solodrai <ihor.solodrai@linux.dev>