ci/diffs: add patches with vmalloc fixes
Recent mm patch [1] causes a regression in BPF verifier which leads to
selftest failures. In particular verifier_loops1 test gets killed by
the test_progs watchdog [2], although normally it passes in a few
seconds.

This has been reported on the mailing list [3], and the fixes are in
flight [4]. Add them as CI-specific patches for now.

[1] https://lore.kernel.org/all/20250424023119.work.333-kees@kernel.org
[2] https://github.com/kernel-patches/bpf/actions/runs/15144596086/job/42577510982
[3] https://lore.kernel.org/lkml/20250515-bpf-verifier-slowdown-vwo2meju4cgp2su5ckj@6gi6ssxbnfqg/
[4] https://lore.kernel.org/all/20250515214020.work.519-kees@kernel.org/

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>